Started Laser Subclass of Attack Object & Image_Draw_List Vector
Attack Object is made to be overwritten by subclasses Laser and Bomb.
Significant progress made on Laser subclass.  None for bomb.  Get_Shape
method for player refactored to return a pointer to the rectangle shape
the image is mapped to, so that we can have a vector holding pointers to
every rectangleshape we want to draw.